+++ src/rasterizer_clipping.cpp
@@ -528,7 +528,7 @@ void PixelShader(const Pixel& p) {
 	float radius = distanceBetweenPoints(pos3d, lightPos);
 
 	//The power per area at this point
-	vec3 B = lightPower / (4 * pi * pow(radius,3));
+	vec3 B = lightPower / (4 * pi * (float) pow(radius,3));
 
 	//unit vector describing normal of surface
 	vec3 n = currentNormal;
